% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTVsuggest.R
\name{CTVsuggest}
\alias{CTVsuggest}
\title{Output CRAN Task View suggestions}
\usage{
CTVsuggest(taskview = "Econometrics", n = 5, ignore = NULL, package = NA)
}
\arguments{
\item{taskview}{A character vector with one element, must be one of the \href{https://github.com/cran-task-views/ctv#available-task-views}{Task Views available}}

\item{n}{An integer that decides the number of suggestions to show.}

\item{ignore}{A character vector of package names that you want to ignore from output suggestions.}
}
\value{
A \link{data.frame} with suggested packages and there classification probability.
}
\description{
The \code{CTVsuggest()} function takes a Task View name and argument \code{n},
then outputs a \link{data.frame} containing the top \code{n} recommendations for the chosen Task View.
}
\details{
The predicted probabilities are computed from the model object constructed with the \href{https://dylandijk.github.io/CTVsuggestTrain/reference/Train_model.html}{\code{CTVsuggestTrain::Train_model()}} function.
}
\examples{
# Outputting top 5 suggestions for the Econometrics Task View,
# whilst ignoring the GVARX package, that is GVARX will not appear in the suggestions.
CTVsuggest(taskview = "Econometrics",
           n = 5, ignore = "GVARX")

#

}
