% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTVsuggest.R
\name{CTVsuggest}
\alias{CTVsuggest}
\title{Output CRAN Task View suggestions}
\usage{
CTVsuggest(taskview = "Econometrics", n = 5)
}
\arguments{
\item{taskview}{A character vector with one element, must be one of the \href{https://github.com/cran-task-views/ctv#available-task-views}{Task Views available}}

\item{n}{An integer that decides the number of suggestions to show.}
}
\value{
A \link{data.frame} with suggested packages and there classification probability.
}
\description{
The \code{CTVsuggest()} function takes a Task View name and argument \code{n},
then outputs a \link{data.frame} containing the top \code{n} recommendations for the chosen Task View.
The predicted probabilities are computed from the model object constructed with the \code{\link[CTVsuggestTrain:Train_model]{CTVsuggestTrain::Train_model()}} function.
}
\examples{
\donttest{
CTVsuggest(taskview = "Econometrics", n = 5)
}


}
